Fixed unit tests that had broken due to various code changes.  Moved a couple tests suites from the network server project to the directory server project.
