+++ directoryServer/pom.xml
@@ -215,6 +215,13 @@
         </dependency>
         -->
 
+        <dependency>
+            <groupId>org.json</groupId>
+            <artifactId>json</artifactId>
+            <version>20160212</version>
+            <scope>test</scope>
+        </dependency>
+
 
 
     </dependencies>
