+++ directoryServer/src/main/java/org/pesc/service/EndpointService.java
@@ -4,12 +4,14 @@ import com.sun.org.apache.xpath.internal.operations.Bool;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.pesc.api.StringUtils;
+import org.pesc.api.model.AuthUser;
 import org.pesc.api.model.Endpoint;
 import org.pesc.api.model.Organization;
 import org.pesc.api.repository.EndpointRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PostAuthorize;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
