+++ directoryServer/src/main/resources/config/application-qa.properties
@@ -93,3 +93,4 @@ url.messages=https://${directory.server}/admin#/messages
 url.organization=https://${directory.server}/admin#/organization/
 url.user=https://${directory.server}/admin#/user/
 email.from=edexchange@ccctechcenter.org
+directory.server.base.url=http://${directory.server}:${http.port}
