+++ directoryServer/src/test/resources/localhost-certificate.pem
@@ -1,0 +1,25 @@
+-----BEGIN CERTIFICATE-----
+MIIEHzCCAwegAwIBAgIJAJ7ZF+rHcGNKMA0GCSqGSIb3DQEBCwUAMIGlMQswCQYD
+VQQGEwJVUzETMBEGA1UECAwKQ2FsaWZvcm5pYTERMA8GA1UEBwwIT3JvdmlsbGUx
+GDAWBgNVBAoMD0NDQyBUZWNoIENlbnRlcjETMBEGA1UECwwKRURFeGNoYW5nZTES
+MBAGA1UEAwwJbG9jYWxob3N0MSswKQYJKoZIhvcNAQkBFhxqd2hldHN0b25lQGNj
+Y3RlY2hjZW50ZXIub3JnMB4XDTE2MDYyNDE0MzU1OVoXDTI2MDYyMjE0MzU1OVow
+gaUxCzAJBgNVBAYTAlVTMRMwEQYDVQQIDApDYWxpZm9ybmlhMREwDwYDVQQHDAhP
+cm92aWxsZTEYMBYGA1UECgwPQ0NDIFRlY2ggQ2VudGVyMRMwEQYDVQQLDApFREV4
+Y2hhbmdlMRIwEAYDVQQDDAlsb2NhbGhvc3QxKzApBgkqhkiG9w0BCQEWHGp3aGV0
+c3RvbmVAY2NjdGVjaGNlbnRlci5vcmcwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAw
+ggEKAoIBAQCpwrXOBGlLl8tKsFsDWfXGIylRRon+cTg7U343Z5yp24ViYPKoApJp
+cbs33frGigmk0F6ey/+MJ4960H8OwF0qiwx6bJqyPZeaHCgyTP1U1wPiPoa1ssTV
+O/ZlvCgzTpMrf1gU/gsnDPTbE1aqP2wHoWOiFUNlf0nP19kIsF9aMiFSXBDFtqym
+uuGl5OBXUqpASkM1sLrjI/P2E1RsXUzSMDzdeK/ifNgUywbAjPHacNUhoqaYinrI
+rIrG6D8Y+NbWoi37qu0QJXRaw+9XcFVQfekIPJKVn3k6QcRsJ+jIN6Mf905mRnV8
+xdd/QN0gtTVDjyPQjyQ9G46EGg1cgKWtAgMBAAGjUDBOMB0GA1UdDgQWBBTKlp24
+DLfxRx6vQUjsrPBzYJf3IjAfBgNVHSMEGDAWgBTKlp24DLfxRx6vQUjsrPBzYJf3
+IjAMBgNVHRMEBTADAQH/MA0GCSqGSIb3DQEBCwUAA4IBAQBOiKQnJnR9EU0jzu/X
+Ougz1MYTSp8KJSAbQUdg56ZdqnDtN8r3IrUiCRctOtjc0+03i4bcuF/Aw1CEbHZN
+Vn7gfF8gFlzj2TnPs8FM3uzvlslYVrRDE/ttCKxT/RmZTgN/OfLyEs/5P13uzUj7
+BKL1sqOi2//s2bQNmPJ0FDmMQOYbEAD/K/LarIpNnzlMG24w+Qo4EErg8iZwcYeg
+OGWHjda1HACcTXWNc23YDFziHJHPERHmPx1QUBW32hsGOBz3r/eOb/QDSlnHTEra
+D/X4jUKHL69DTFCC7C6/+xYEdrdnft+21cdyetJm8G1c6CihGaXCHaRHGMfuZ7u7
+8SBf
+-----END CERTIFICATE-----
