+++ networkServer/src/main/resources/config/application-dev.properties
@@ -14,7 +14,7 @@ db.url=jdbc:mysql://${db.server}:${db.port}/pesc_networkserver
 db.username=ccctc
 db.password=ccc.dev
 
-directory.server.base.url=http://localhost:8080
+directory.server.base.url=http://localhost
 
 # network server properties
 networkServer.hostname=localhost
